refactor(linter): shared walkValues primitive, range helpers, import-fix machinery
